me.hfox => uk.hfox
