+++ morphix-api/src/main/java/uk/hfox/morphix/query/FieldQuery.java
@@ -16,7 +16,7 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  * ========================LICENSE_END========================
  */
-package me.hfox.morphix.query;
+package uk.hfox.morphix.query;
 
 /**
  * Created by Harry on 28/11/2017.
@@ -41,7 +41,7 @@ public interface FieldQuery<T> {
 
     /**
      * Set if this field is an inverse query
-     * @see me.hfox.morphix.query.FieldQuery#not()
+     * @see FieldQuery#not()
      * @param not true if this should be an inverse query, otherwise false
      * @return The current query with the not set
      */
